BackEnd/requirements.txt
asgiref
certifi
chardet
coreapi
coreschema
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
idna
itypes
Jinja2
MarkupSafe
Pillow
PyJWT
pytz
requests
sqlparse
uritemplate
urllib3
drf-yasg
django-autoslug
django-webpack-loader